+++ src/Controllers/authCtrl.ts
@@ -1,11 +1,9 @@
-import { PrismaClient } from "@prisma/client";
-import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 import { NextFunction, Request, Response } from "express";
 import { HTTP_STATUS } from "../constants/httpStatusCode";
 import ClientError from "../errors/clientError";
 import { AuthService } from "../services/authService";
+import handleRegistrationError from "../utils/handleRegistrationError";
 import { sendResponse } from "../utils/sendResponse";
-
 
 export const registerCtrl = async (
   req: Request,
@@ -50,23 +48,4 @@ export const googleLoginCtrl = async (
   } catch (error) {
     handleRegistrationError(error, res, next);
   }
-};
-
-const handleRegistrationError = (
-  error: any,
-  res: Response,
-  next: NextFunction
-) => {
-  if (error instanceof PrismaClientKnownRequestError) {
-    console.error("Error conocido de Prisma:", error.message);
-    return res.status(500).json({ error: "Error al procesar la solicitud." });
-  }
-
-  if (error instanceof ClientError) {
-    console.error("Error personalizado:", error.message);
-    return next(error);
-  }
-
-  console.error("Error desconocido:", error.message);
-  return res.status(500).json({ error: "Error al procesar la solicitud." });
 };
